fix(levelup): replace table connection in Pitch
replace @OneToOne by @OneToMany related to review table
