+++ src/main/java/com/greenfoxacademy/levelup/model/User.java
@@ -1,0 +1,35 @@
+package com.greenfoxacademy.levelup.model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import javax.persistence.MapsId;
+import javax.persistence.OneToMany;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+public class Person {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long id;
+  private String name;
+  private String email;
+  private String role;
+  @ManyToOne
+  private BadgeLevel badgeLevel;
+  @OneToMany
+  @MapsId
+  private Review review;
+  @OneToMany
+  @MapsId
+  private Pitch pitch;
+}
