+++ src/main/java/com/greenfoxacademy/levelup/repository/IUserRepository.java
@@ -1,11 +1,11 @@
 package com.greenfoxacademy.levelup.repository;
 
-import com.greenfoxacademy.levelup.model.Person;
+import com.greenfoxacademy.levelup.model.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPersonRepository extends CrudRepository<Person, Long> {
+public interface IPersonRepository extends CrudRepository<User, Long> {
 
 }
